Rewrites internal module loading
( to make bundle size smaller )

This feature rewrites shiny so we can referer to the src/index.js even without having a working treeshaking going. Because it seems it is very hard to get to work, I think impossible if you are not on webpack 4. Instead babel transform all the ref. to the file they are coming from.

Unfortunally this means adding "import/prefer-default-export": 0, since we need to rename import to different files.

See https://material-ui.com/guides/minimizing-bundle-size/ for how matarial-ui has done it.
We should keep an eye on how they fix it in the future, but now, I think this is the best approche.