fix(LinkBarElementBase): counteract labrador form styles
At the time of this commit, we are releasing the seher header
from wolverine as an edge side include in seher pages served
by labrador. In the seher styleguide of old times, there are
some reset styles that are more specific than the styles we
apply to the search field in this header.

This commit increases the specificity of LinkBarElementBase
to try to avoid our styles being overridden by the old
styleguide.