fix(Dropdown): Account for emotion 10 JIT style tags
In emotion 10, style tags will be inserted just in time. That means that
they might be inserted just before this component. Which again means
that this component no longer will be an adjacent sibling to the
dropdown button in LinkBarDropdown in our most common use case.

Therefore, we switch to the general sibling selector.